+++ bricks/basic-bricks/src/utils.ts
@@ -4,7 +4,7 @@ import { BrickConf, RouteConfOfBricks } from "@next-core/brick-types";
 export function getNewUIStatus(isSetPageTitleStyle?: boolean): boolean {
   const { getFeatureFlags, getCurrentRoute } = getRuntime();
   const featureFlag = getFeatureFlags()["support-ui-8.0-base-layout"];
-  const { bricks } = getCurrentRoute() as RouteConfOfBricks;
+  const { bricks = [] } = (getCurrentRoute() ?? {}) as RouteConfOfBricks;
   const tplNames = [
     "base-layout.tpl-base-page-module",
     "base-layout.tpl-base-page-module-cmdb",