+++ bricks/forms/src/cmdb-cascader/index.spec.ts
@@ -26,7 +26,7 @@ describe("forms.cmdb-cascader", () => {
     await jest.runAllTimers();
     element.addEventListener("forms.cmdb-cascader.change", fn);
     element.handleChange("xxx");
-    expect(fn).toBeCalled();
+    expect(fn).toBeCalledTimes(0);
   });
 
   it("handleChangeV2 event", async () => {
@@ -37,6 +37,6 @@ describe("forms.cmdb-cascader", () => {
     await jest.runAllTimers();
     element.addEventListener("forms.cmdb-cascader.change.v2", fn);
     element.handleChangeV2("xxx");
-    expect(fn).toBeCalled();
+    expect(fn).toBeCalledTimes(1);
   });
 });