+++ bricks/forms/src/general-select/GeneralSelect.tsx
@@ -15,7 +15,7 @@ import {
   GeneralComplexOption,
 } from "@next-libs/forms";
 import style from "./GeneralSelect.module.css";
-import { debounce, groupBy, isNil, isEqual, trim } from "lodash";
+import { debounce, groupBy, isNil, isEqual, trim, keyBy } from "lodash";
 import { GeneralOption } from "@next-libs/forms/dist/types/interfaces";
 import { maxTagCountType } from "./index";
 export const setTooltip = (event: React.MouseEvent) => {
@@ -46,6 +46,8 @@ const applyArgs = (args: any[] | ((query: string) => any[]), query: string) => {
   return args;
 };
 
+type RequestStatus = "loading" | "success" | "error";
+
 export interface GeneralSelectProps extends FormItemWrapperProps {
   options: GeneralComplexOption[];
   fields?: Partial<GeneralComplexOption>;
@@ -57,7 +59,9 @@ export interface GeneralSelectProps extends FormItemWrapperProps {
   dropdownMatchSelectWidth?: boolean;
   onChange?: (value: any, options: GeneralComplexOption[]) => void;
   onChangeV2?: (value: any) => void;
-  onOptionDataChange?: (data: GeneralComplexOption) => void;
+  onOptionDataChange?: (
+    data: GeneralComplexOption | GeneralComplexOption[]
+  ) => void;
   allowClear?: boolean;
   showSearch?: boolean;
   disabled?: boolean;
@@ -72,6 +76,7 @@ export interface GeneralSelectProps extends FormItemWrapperProps {
   onSearch?: (value: string) => void;
   useBackend?: UseBackendConf & {
     onValueChangeArgs?: any[] | ((...args: any[]) => any[]);
+    emptyConfig?: Partial<Record<RequestStatus, EasyopsEmptyProps>>;
   };
   onDebounceSearch?: (value: string) => void;
   debounceSearchDelay?: number;
@@ -109,9 +114,9 @@ export function GeneralSelectLegacy(
   } = props;
   const [checkedValue, setCheckedValue] = useState(props.value);
   const [options, setOptions] = useState<GeneralComplexOption[]>(props.options);
-  const [loading, setLoading] = useState(false);
+  const [requestStatus, setRequestStatus] = useState<RequestStatus>();
   const shouldTriggerOnValueChangeArgs = useRef(true);
-  const curOptionData = useRef<GeneralComplexOption>();
+  const curOptionData = useRef<GeneralComplexOption | GeneralComplexOption[]>();
   const request = useProvider({ cache: false });
   React.useEffect(() => {
     if (suffixBrick) {
@@ -132,8 +137,18 @@ export function GeneralSelectLegacy(
   }, [props.options]);
 
   useEffect(() => {
-    if (Array.isArray(checkedValue)) {
-      // Todo(nlicro) 多选支持该事件
+    if (props.mode === "multiple") {
+      const optionDataMap = keyBy(options, "value");
+      const preOptionDataMap = keyBy(curOptionData.current, "value");
+
+      const newOptionsData = []
+        .concat(checkedValue ?? [])
+        .map((v) => optionDataMap[v] || preOptionDataMap[v]);
+
+      if (!isEqual(curOptionData.current, newOptionsData)) {
+        curOptionData.current = newOptionsData;
+        onOptionDataChange?.(newOptionsData);
+      }
       return;
     } else {
       const newOptionData = options?.find((v) => v.value === checkedValue);
@@ -185,7 +200,7 @@ export function GeneralSelectLegacy(
           } = props.useBackend;
           (async () => {
             try {
-              setLoading(true);
+              setRequestStatus("loading");
               const actualArgs = applyArgs(
                 type === "search" ? args : onValueChangeArgs,
                 value
@@ -197,11 +212,11 @@ export function GeneralSelectLegacy(
                 transformedData as unknown as GeneralOption[],
                 props.fields as any
               );
+              setRequestStatus("success");
               setOptions(actualData);
             } catch (e) {
+              setRequestStatus("error");
               handleHttpError(e);
-            } finally {
-              setLoading(false);
             }
           })();
         } else {
@@ -291,6 +306,14 @@ export function GeneralSelectLegacy(
     ));
   };
 
+  const notFoundContent = useMemo(() => {
+    const _emptyProps =
+      (isSearchable(props.useBackend) &&
+        props.useBackend.emptyConfig?.[requestStatus]) ||
+      emptyProps;
+    return <EasyopsEmpty {..._emptyProps} />;
+  }, [emptyProps, requestStatus, props.useBackend]);
+
   return (
     <Select
       ref={ref}
@@ -314,8 +337,8 @@ export function GeneralSelectLegacy(
         ? { getPopupContainer: (triggerNode) => triggerNode.parentElement }
         : {})}
       dropdownStyle={{ padding: "2px", ...props.dropdownStyle }}
-      notFoundContent={<EasyopsEmpty {...emptyProps} />}
-      loading={loading}
+      notFoundContent={notFoundContent}
+      loading={requestStatus === "loading"}
       bordered={props.bordered}
       onFocus={() => {
         props.onFocus?.();