+++ bricks/flow-builder/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.25.0](https://github.com/easyops-cn/next-basics/compare/@next-bricks/flow-builder@1.24.11...@next-bricks/flow-builder@1.25.0) (2023-02-08)
+
+
+### Features
+
+* add processor ([3dde02a](https://github.com/easyops-cn/next-basics/commit/3dde02a1140d254c844fdb424dc3166829dc41a4))
+
+
+
+
+
 ## [1.24.11](https://github.com/easyops-cn/next-basics/compare/@next-bricks/flow-builder@1.24.10...@next-bricks/flow-builder@1.24.11) (2023-02-01)